+++ src/shared/settings.ts
@@ -7,6 +7,12 @@ export interface Settings {
     inputBoxMaxWidthStyle: string;
     textColorUserStyle: string;
     textColorNonUserStyle: string;
+    customBackgroundsEnabled: boolean;
+    syncBackgroundColors: boolean;
+    conversationBackgroundStyle: string;
+    sidebarBackgroundStyle: string;
+    syncedBackgroundStyle: string;
+    scrollToTopEnabled: boolean;
 }
 
 export const defaultSettings: Settings = {
@@ -18,4 +24,10 @@ export const defaultSettings: Settings = {
     inputBoxMaxWidthStyle: "94",
     textColorUserStyle: "#FFFFFF",
     textColorNonUserStyle: "#FFFFFF",
+    customBackgroundsEnabled: false,
+    syncBackgroundColors: false,
+    conversationBackgroundStyle: "#212121",
+    sidebarBackgroundStyle: "#171717",
+    syncedBackgroundStyle: "#212121",
+    scrollToTopEnabled: true,
 };